+++ utility/test/Test_string_to_array.cpp
@@ -44,9 +44,9 @@ TEST(string_to_array_d, numeric_and_non_numeric)
     // Test comma and space separated values with additional non-numeric fields.
     double list[NUM_DOUBLES];
     char line[] = "hello 1.0,2.0 3.0, there,4.0     5.0 6.0";
-    int filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
-    ASSERT_EQ(6, filled);
-    for (int i = 0; i < filled; ++i)
+    size_t filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
+    ASSERT_EQ((size_t)6, filled);
+    for (size_t i = 0; i < filled; ++i)
         EXPECT_DOUBLE_EQ((double)(i+1), list[i]);
 }
 
@@ -55,8 +55,8 @@ TEST(string_to_array_d, empty_string)
     // Test empty string.
     double list[NUM_DOUBLES];
     char line[] = "";
-    int filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
-    ASSERT_EQ(0, filled);
+    size_t filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
+    ASSERT_EQ((size_t)0, filled);
 }
 
 TEST(string_to_array_d, single_space)
@@ -64,8 +64,8 @@ TEST(string_to_array_d, single_space)
     // Test single space.
     double list[NUM_DOUBLES];
     char line[] = " ";
-    int filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
-    ASSERT_EQ(0, filled);
+    size_t filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
+    ASSERT_EQ((size_t)0, filled);
 }
 
 TEST(string_to_array_d, negative_integers)
@@ -73,9 +73,9 @@ TEST(string_to_array_d, negative_integers)
     // Test negative integers.
     double list[NUM_DOUBLES];
     char line[] = "-4,-3,-2 -1 0";
-    int filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
-    ASSERT_EQ(5, filled);
-    for (int i = 0; i < filled; ++i)
+    size_t filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
+    ASSERT_EQ((size_t)5, filled);
+    for (int i = 0; i < (int)filled; ++i)
         EXPECT_DOUBLE_EQ((double)(i-4), list[i]);
 }
 
@@ -84,8 +84,8 @@ TEST(string_to_array_d, non_matching_string)
     // Test non-matching string.
     double list[NUM_DOUBLES];
     char line[] = "nobody home";
-    int filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
-    ASSERT_EQ(0, filled);
+    size_t filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
+    ASSERT_EQ((size_t)0, filled);
 }
 
 TEST(string_to_array_d, too_many_items)
@@ -93,9 +93,9 @@ TEST(string_to_array_d, too_many_items)
     // Test too many items.
     double list[NUM_DOUBLES];
     char line[] = "0.1 0.2 0.3   ,  0.4 0.5 0.6 0.7 0.8 0.9 1.0";
-    int filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
-    ASSERT_EQ(NUM_DOUBLES, filled);
-    for (int i = 0; i < filled; ++i)
+    size_t filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
+    ASSERT_EQ((size_t)NUM_DOUBLES, filled);
+    for (size_t i = 0; i < filled; ++i)
         EXPECT_DOUBLE_EQ((i+1)/10.0, list[i]);
 }
 
@@ -104,8 +104,8 @@ TEST(string_to_array_d, single_item)
     // Test single item.
     char line[] = "   0.1 ";
     double par;
-    int filled = oskar_string_to_array_d(line, 1, &par);
-    ASSERT_EQ(1, filled);
+    size_t filled = oskar_string_to_array_d(line, 1, &par);
+    ASSERT_EQ((size_t)1, filled);
     EXPECT_DOUBLE_EQ(0.1, par);
 }
 
@@ -114,8 +114,8 @@ TEST(string_to_array_d, comment_line)
     // Test comment line.
     double list[NUM_DOUBLES];
     char line[] = "# This is a comment.";
-    int filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
-    ASSERT_EQ(0, filled);
+    size_t filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
+    ASSERT_EQ((size_t)0, filled);
 }
 
 TEST(string_to_array_d, comment_line_with_space)
@@ -123,8 +123,8 @@ TEST(string_to_array_d, comment_line_with_space)
     // Test comment line with preceding space.
     double list[NUM_DOUBLES];
     char line[] = " # This is another comment.";
-    int filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
-    ASSERT_EQ(0, filled);
+    size_t filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
+    ASSERT_EQ((size_t)0, filled);
 }
 
 TEST(string_to_array_d, comment_at_end)
@@ -132,9 +132,9 @@ TEST(string_to_array_d, comment_at_end)
     // Test line with comment at end.
     double list[NUM_DOUBLES];
     char line[] = " 1.0 1.1 1.2 1.3 # This is another comment.";
-    int filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
-    ASSERT_EQ(4, filled);
-    for (int i = 0; i < filled; ++i)
+    size_t filled = oskar_string_to_array_d(line, NUM_DOUBLES, list);
+    ASSERT_EQ((size_t)4, filled);
+    for (size_t i = 0; i < filled; ++i)
         EXPECT_DOUBLE_EQ((double)(i/10.0 + 1), list[i]);
 }
 
@@ -145,8 +145,8 @@ TEST(string_to_array_s, empty_string)
     // Test empty string.
     char *list[NUM_STRINGS];
     char line[] = "";
-    int filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
-    ASSERT_EQ(0, filled);
+    size_t filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
+    ASSERT_EQ((size_t)0, filled);
 }
 
 TEST(string_to_array_s, single_space)
@@ -154,8 +154,8 @@ TEST(string_to_array_s, single_space)
     // Test single space.
     char *list[NUM_STRINGS];
     char line[] = " ";
-    int filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
-    ASSERT_EQ(0, filled);
+    size_t filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
+    ASSERT_EQ((size_t)0, filled);
 }
 
 TEST(string_to_array_s, normal_line)
@@ -163,8 +163,8 @@ TEST(string_to_array_s, normal_line)
     // Test normal use case.
     char *list[NUM_STRINGS];
     char line[] = "*, *, 10, 20, AZEL";
-    int filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
-    ASSERT_EQ(5, filled);
+    size_t filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
+    ASSERT_EQ((size_t)5, filled);
     EXPECT_STREQ("*",    list[0]);
     EXPECT_STREQ("*",    list[1]);
     EXPECT_STREQ("10",   list[2]);
@@ -180,8 +180,8 @@ TEST(string_to_array_s, comment_line)
     // Test comment line.
     char *list[NUM_STRINGS];
     char line[] = "# This is a comment.";
-    int filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
-    ASSERT_EQ(0, filled);
+    size_t filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
+    ASSERT_EQ((size_t)0, filled);
 }
 
 TEST(string_to_array_s, comment_line_with_space)
@@ -189,8 +189,8 @@ TEST(string_to_array_s, comment_line_with_space)
     // Test comment line with preceding space.
     char *list[NUM_STRINGS];
     char line[] = " # This is another comment.";
-    int filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
-    ASSERT_EQ(0, filled);
+    size_t filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
+    ASSERT_EQ((size_t)0, filled);
 }
 
 TEST(string_to_array_s, comment_at_end)
@@ -198,8 +198,8 @@ TEST(string_to_array_s, comment_at_end)
     // Test line with comment at end.
     char *list[NUM_STRINGS];
     char line[] = " 1.0 1.1 1.2 1.3 # This is another comment.";
-    int filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
-    ASSERT_EQ(4, filled);
+    size_t filled = oskar_string_to_array_s(line, NUM_STRINGS, list);
+    ASSERT_EQ((size_t)4, filled);
     EXPECT_STREQ("1.0", list[0]);
     EXPECT_STREQ("1.1", list[1]);
     EXPECT_STREQ("1.2", list[2]);
@@ -212,27 +212,27 @@ TEST(string_to_array_realloc_s, reuse_buffer)
 {
     // Test lines with comments and blanks.
     char **list = 0;
-    int n = 0, filled = 0;
+    size_t n = 0, filled = 0;
 
     {
         char line[] = "# This is a comment.";
         filled = oskar_string_to_array_realloc_s(line, &n, &list);
-        ASSERT_EQ(0, filled);
-        ASSERT_EQ(0, n);
+        ASSERT_EQ((size_t)0, filled);
+        ASSERT_EQ((size_t)0, n);
     }
 
     {
         char line[] = " # This is another comment.";
         filled = oskar_string_to_array_realloc_s(line, &n, &list);
-        ASSERT_EQ(0, filled);
-        ASSERT_EQ(0, n);
+        ASSERT_EQ((size_t)0, filled);
+        ASSERT_EQ((size_t)0, n);
     }
 
     {
         char line[] = "1, *, 10, 20, AZEL";
         filled = oskar_string_to_array_realloc_s(line, &n, &list);
-        ASSERT_EQ(5, filled);
-        ASSERT_EQ(5, n);
+        ASSERT_EQ((size_t)5, filled);
+        ASSERT_EQ((size_t)5, n);
         EXPECT_STREQ("1",    list[0]);
         EXPECT_STREQ("*",    list[1]);
         EXPECT_STREQ("10",   list[2]);
@@ -246,15 +246,15 @@ TEST(string_to_array_realloc_s, reuse_buffer)
     {
         char line[] = " ";
         filled = oskar_string_to_array_realloc_s(line, &n, &list);
-        ASSERT_EQ(0, filled);
-        ASSERT_EQ(5, n);
+        ASSERT_EQ((size_t)0, filled);
+        ASSERT_EQ((size_t)5, n);
     }
 
     {
         char line[] = "2, 0, 3, 34.5, 67.8, RADEC";
         filled = oskar_string_to_array_realloc_s(line, &n, &list);
-        ASSERT_EQ(6, filled);
-        ASSERT_EQ(6, n);
+        ASSERT_EQ((size_t)6, filled);
+        ASSERT_EQ((size_t)6, n);
         EXPECT_STREQ("2",     list[0]);
         EXPECT_STREQ("0",     list[1]);
         EXPECT_STREQ("3",     list[2]);
@@ -270,8 +270,8 @@ TEST(string_to_array_realloc_s, reuse_buffer)
     {
         char line[] = "1, 2, 50, 60, AZEL # Another comment";
         filled = oskar_string_to_array_realloc_s(line, &n, &list);
-        ASSERT_EQ(5, filled);
-        ASSERT_EQ(6, n);
+        ASSERT_EQ((size_t)5, filled);
+        ASSERT_EQ((size_t)6, n);
         EXPECT_STREQ("1",    list[0]);
         EXPECT_STREQ("2",    list[1]);
         EXPECT_STREQ("50",   list[2]);
